Load model with Task and add DataPipeline into DataModule as lightning-flash
